requirements.txt
streamlit
spacy
sentence-transformers
google-generativeai
python-dotenv
pdfplumber
plotly
st-annotated-text
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1.tar.gz
